+++ src/forfiles/__init__.py
@@ -1,13 +1,14 @@
 """API entry point for the forfiles package."""
 
-from .fs import dir_action, dir_create, dir_delete, filter_type
+from .fs import dir_create, dir_delete, filter_type, iterate_files, process_files
 from .image import resize, scale, to_png
 
 __all__ = [
-    'dir_action',
     'dir_create',
     'dir_delete',
     'filter_type',
+    'iterate_files',
+    'process_files',
     'resize',
     'scale',
     'to_png',