+++ src/forfiles/fs.py
@@ -1,7 +1,7 @@
 """Tools for the filesystem."""
 
 import os
-from collections.abc import Callable
+from collections.abc import Callable, Generator
 from pathlib import Path
 from shutil import rmtree
 from typing import Concatenate, ParamSpec
@@ -66,13 +66,15 @@ def dir_delete(directory: StrOrBytesPath) -> None:
 P = ParamSpec('P')
 
 
-def dir_action(
+def process_files(
     directory: StrOrBytesPath,
     fn: Callable[Concatenate[Path, P], None],
     *args: P.args,
     **kwargs: P.kwargs,
 ) -> None:
     """Iterate through a directory and execute a function for each file in the directory.
+
+    You can use `iterate_files` if you prefer to iterate through files with a generator.
 
     Args:
         directory (StrOrBytesPath):
@@ -90,16 +92,40 @@ def dir_action(
     Examples:
         The following code demonstrates how to print the contents of a directory:
 
-        >>> def print_file_contents(file_path):
+        >>> def print_contents(file_path):
         ...     with open(file_path, 'r') as file:
         ...         print(file.read())
 
-        >>> dir_action('/path/to/directory', print_file_contents)
-
-        The example prints the contents of each file in the specified directory.
+        >>> process_files('/path/to/directory', print_contents)
 
     """
     directory = process_path(directory) if not isinstance(directory, Path) else directory
     for file_path in directory.rglob('*'):
         if file_path.is_file():
             fn(file_path, *args, **kwargs)
+
+
+def iterate_files(directory: StrOrBytesPath) -> Generator[Path, None, None]:
+    """Iterate through a directory and yield the paths of each file.
+
+    You can use `process_files` if you prefer to process files with a callback function.
+
+    Args:
+        directory (StrOrBytesPath):
+            Path of the directory to iterate through.
+
+    Yields:
+        Path:
+            The path of each file in the directory.
+
+    Examples:
+        The following code demonstrates how to print the paths of all files:
+
+        >>> for file in iterate_files('/path/to/directory'):
+        >>>     print(file)
+
+    """
+    directory = process_path(directory) if not isinstance(directory, Path) else directory
+    for file_path in directory.rglob('*'):
+        if file_path.is_file():
+            yield file_path